refactor: organizar linhas das lógicas, comentar as mesmas e separar os estilos do arquivo
